refactor(user): Refactor User domain in user-api
Change column name of primary key in user table from user_id to id.
Change length of password attribute and sequence of configuration components of phone_number attribute.
Remove constructors with no parameter of several classes which is not an entity.
Change AccessLevel of constructor with no parameter to 'protected' in User entity.